Add wrapper for polkit_system_bus_name_get_user_sync
This adds a new method to SystemBusNameSubject to get to the
UnixUserIdentity for the subject.

This feature depends on new functionality added in polkit 0.113, the
dependency is nevertheless not raised as the new API is backported in
distribution packages and we don't want to disrupt developer workflows
by depending on a version which is not available yet.

REVIEW: 126723