Port to Qt6
But still keep the option to build against Qt5 for now, because one
might have ported and unported consumers at the same time. This was
similarly done with the move from Qt4 to Qt5.

Instead of having two different templates for pkgconfig files which
only differ in "Requires: Qt{5,6}Foo" let cmake fill in the selected
Qt.